fix user story end task

+++ components/task/task.php
@@ -5,36 +5,38 @@ $query->execute();
 $taskList = $query->fetchAll();
 
 foreach ($taskList as $task) {
-    echo "
-    <ul>
-    <li class=\"main__container\" data-id=\"" . $task['id_task'] . "\">
-    <div class=\"left__container\">
+    if ($task['task_state'] === 0) {   
+        echo "
+        <ul>
+        <li class=\"main__container\" data-id=\"" . $task['id_task'] . "\">
+        <div class=\"left__container\">
         <div class=\"top__container\">
-           <a href=\"components/task/endtask.php?token=" . $_SESSION['token'] . "&id_task=" . $task['id_task'] . "\">☑️</a>
-            <div class=\"top__container--task_name\">
-                <p>" . $task['task_name'] . "</p>
-            </div>
+        <a href=\"components/task/endtask.php?token=" . $_SESSION['token'] . "&id=" . $task['id_task'] . "\">☑️</a>
+        <div class=\"top__container--task_name\">
+        <p>" . $task['task_name'] . "</p>
+        </div>
         </div>
         <div class=\"bot__container\">
-            <ul class=\"task__action\">
-                <li class=\"btn\">
-                <a href=\"components/task/deletetask.php?token=" . $_SESSION['token'] . "&id_task=" . $task['id_task'] . "\">✍🏼</a>   
-                </li>
-                <li class=\"btn\">
-                    <a href=\"components/task/deletetask.php?token=" . $_SESSION['token'] . "&id_task=" . $task['id_task'] . "\">🗑️</a> 
-                </li>
-            </ul>
+        <ul class=\"task__action\">
+        <li class=\"btn\">
+        <a href=\"components/task/deletetask.php?token=" . $_SESSION['token'] . "&id_task=" . $task['id_task'] . "\">✍🏼</a>   
+        </li>
+        <li class=\"btn\">
+        <a href=\"components/task/deletetask.php?token=" . $_SESSION['token'] . "&id_task=" . $task['id_task'] . "\">🗑️</a> 
+        </li>
+        </ul>
         </div>
-    </div>
-    
-    <div class=\"right__container\">
+        </div>
+        
+        <div class=\"right__container\">
         <div class=\"right__container--arrows\">
-            <p><img src=\"../../assets/svg/angles-up-solid.svg\" alt=\"\"></p>
-            <p><img src=\"../../assets/svg/angles-down-solid.svg\" alt=\"\"></p>
+        <p><img src=\"../../assets/svg/angles-up-solid.svg\" alt=\"\"></p>
+        <p><img src=\"../../assets/svg/angles-down-solid.svg\" alt=\"\"></p>
         </div>
-    </div>
-</li>
-</ul>";
+        </div>
+        </li>
+        </ul>";
+    }
 }
-
-?>
+    
+    ?>
